code/system/requirement.txt
h5py
matplotlib
opacus
pandas
seaborn
scikit-learn
scipy
torchaudio
torchtext
torchvision
ujson
mlflow
calmsize
